+++ BasketUI/BasketLogicTests/RulesEngineTests.cs
@@ -13,7 +13,7 @@ namespace BasketLogicTests
         public void RulesEngineHasApplyRulesMethod()
         {
             Mock<IRulesEngine> mockRulesEngine = new Mock<IRulesEngine>();
-            mockRulesEngine.Setup(m => m.ApplyRules(It.IsAny<List<BasketItem>>(), It.IsAny<IList<IRule>>())).Returns(0.0f);
+            mockRulesEngine.Setup(m => m.ApplyRules(It.IsAny<List<BasketItem>>(), It.IsAny<IList<IRule>>())).Returns(0.0m);
         }
 
         [TestMethod]
@@ -21,11 +21,11 @@ namespace BasketLogicTests
         {
             RulesEngine rulesEngine = new RulesEngine();
             Mock<IRule> mockRuleA = new Mock<IRule>();
-            float deductionA = 1.0f;
+            decimal deductionA = 1.0m;
             mockRuleA.Setup(m => m.Apply(It.IsAny<List<BasketItem>>())).Returns(-deductionA);
 
             Mock<IRule> mockRuleB = new Mock<IRule>();
-            float deductionB = 0.7f;
+            decimal deductionB = 0.7m;
             mockRuleB.Setup(m => m.Apply(It.IsAny<List<BasketItem>>())).Returns(-deductionB);
 
             List<BasketItem> basketItems = new List<BasketItem>();
@@ -34,7 +34,7 @@ namespace BasketLogicTests
             rules.Add(mockRuleA.Object);
             rules.Add(mockRuleB.Object);
 
-            float adjustment = rulesEngine.ApplyRules(basketItems, rules);
+            decimal adjustment = rulesEngine.ApplyRules(basketItems, rules);
 
             Assert.AreEqual(-deductionA - deductionB, adjustment);
